First pass at refactoring pricing client to generic
