+++ classes/pricing/aws_rekog_pricing_client.php
@@ -1,0 +1,59 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
+/**
+ * Service client for getting AWS pricing information for the Elastic Transcode Services (ETS).
+ *
+ * @package     local_smartmedia
+ * @author      Tom Dickman <tomdickman@catalyst-au.net>
+ * @copyright   2019 Catalyst IT Australia {@link http://www.catalyst-au.net}
+ * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace local_smartmedia\pricing;
+
+defined('MOODLE_INTERNAL') || die;
+
+global $CFG;
+// Autoload the SDK for AWS service usage.
+require_once($CFG->dirroot . '/local/aws/sdk/aws-autoloader.php');
+
+/**
+ * A client for getting pricing information for AWS Elastic Transcode Services.
+ *
+ * @package     local_smartmedia
+ * @author      Tom Dickman <tomdickman@catalyst-au.net>
+ * @copyright   2019 Catalyst IT Australia {@link http://www.catalyst-au.net}
+ * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+class aws_rekog_pricing_client extends aws_base_pricing_client {
+
+    /**
+     * The ServiceCode for Amazon Rekognition Services.
+     */
+    const SERVICE_CODE = 'AmazonRekognition';
+
+    /**
+     * Get the pricing for a specific transcode location.
+     *
+     * @param string $region the region code of an AmazonETS location to get pricing for.
+     *
+     * @return \local_smartmedia\location_transcode_pricing $locationpricing object containing pricing.
+     */
+    public function get_location_pricing($region) {
+
+    }
+}
